==> Instacart Basket Analysis/02 Scripts/2 Jupyter Fundamentals & Python Data Types.ipynb <==
#01. Importing libraries

# Import libraries
import pandas as pd
import numpy as np
import os

a = 50 + 50
b = 200 - 100
c = 20 * 5

a, b, c

x = 20.8
y = 1.4

x / y

'Note' + 'book'

f = 'Learning ' + 'how ' + 'to ' + 'code ' + 'in ' + 'Python.'
g = 'Installing ' + 'Anaconda, ' + 'NumPy, ' + 'and ' + 'Jupyter.'

print(f, g)

# Bonus Task
laugh = 'ha'

laugh * 1000
