==> Instacart Basket Analysis/02 Scripts/3 Intro to Pandas.ipynb <==
# Import libraries
import pandas as pd
import numpy as np
import os

path = r'C:\Users\nukis\Documents\Instacart Basket Analysis'

vars_list = ['order_id', 'user_id', 'order_number', 'order_dow', 'order_hour_of_day', 'days_since_prior_order']

df = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), usecols = vars_list)

df.head(5)

df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'))

df_prods.head(20)

df_prods.tail(35)

df_prods.columns

df_prods.shape

df_prods.dtypes

df_prods.describe()

# Answer number 11
# a. What’s the maximum value of the “aisle_id” column? 134
# b. What’s the data type of the “department_id” column? integer
