==> Instacart Basket Analysis/02 Scripts/4 Data Wrangling & Subsetting.ipynb <==
# Importing libraries

import pandas as pd
import numpy as np
import os

#### Defining path

path = r'C:\Users\nukis\Documents\Instacart Basket Analysis'

#### Importing orders.csv, products.csv and departments.csv

df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), index_col = False)

df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

df_dep = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'departments.csv'), index_col = False)

# 2. Data wrangling steps

#### Q2. Find another identifier variable in the df_ords dataframe that doesn’t need to be included in your analysis as a numeric variable and change it to a suitable format.

# Checking variable types
df_ords.dtypes

# Only 'eval_set' is found needed to be excluded and saved it in the new variable
df_ords_2 = df_ords.drop(columns = ['eval_set'])

df_ords_2

#### Q3. Look for a variable in your df_ords dataframe with an unintuitive name and change its name without overwriting the data frame.

df_ords_2.rename(columns = {'order_dow' : 'order_day_of_week'}, inplace = True)

df_ords_2

#### Q4. Your client wants to know what the busiest hour is for placing orders. Find the frequency of the corresponding variable and share your findings.

# Counting order frequency based on time (hour)
df_ords_2['order_hour_of_day'].value_counts(dropna = False)

##### Busiest hour is at 10:00, hour with highest number of order

df_ords_2.describe()

#### Q5. Determine the meaning behind a value of 4 in the "department_id" column within the df_prods dataframe using a data dictionary.

df_prods.head()

# Checking departments.csv
df_dep.head()

# Transposing df_dep and save into new variable
df_dep_t = df_dep.T

df_dep_t

# Resetting df_dep_t index
df_dep_t.reset_index()

# Taking df_dep_t[0] as new_header
new_header = df_dep_t.iloc[0]

new_header

# Copying 1st row onward into new variable
df_dep_t_new = df_dep_t[1:]

# Set new_header AS new header
df_dep_t_new.columns = new_header

df_dep_t_new

# Converting into data dictionary
df_dep_t_new_dict = df_dep_t_new.to_dict('index')

df_dep_t_new_dict

print(df_dep_t_new_dict['4'])

#### Value of 4 in the "department_id" is produce.

#### Q6. The sales team in your client’s organization wants to know more about "breakfast" item sales. Create a subset containing only the required information.

# This loop step is done in case we dont know where "breakfast", etc. item/department are located.
for key, value in df_dep_t_new_dict.items():
    if value['department'] == 'breakfast' or value['department'] == 'alcohol' or value['department'] == 'deli' or value['department'] == 'beverages' or value['department'] == 'meat seafood': 
        print(key, value)

# Creating subset of breakfast located in department 14
df_prods_breakfast = df_prods[df_prods['department_id'] == 14]

#### Q7. They’d also like to see details about customers who might be throwing dinner parties. Your task is to find all observations from the entire dataframe that include items from the following departments: "alcohol, deli, beverages, and meat/seafood". You’ll need to present this subset to your client.

# Creating combined subset of alcohol, deli, beverages, and meat/seafood located in department 5, 20, 7, 12
df_prods_dinner = df_prods[df_prods['department_id'].isin([5, 20, 7, 12])]

#### Q8. It’s important that you keep track of total counts in your dataframes. How many rows does the last dataframe you created have?

df_prods_dinner.shape

#### Q9. Someone from the data engineers team in Instacart thinks they’ve spotted something strange about the customer with a "user_id" of “1.” Extract all the information you can about this user.

# Extracting user_id = 1
df_user_id_1 = df_ords[df_ords['user_id'] == 1]

df_user_id_1

#### Q10. You also need to provide some details about this user’s behavior. What basic stats can you provide based on the information you have?

# Behavior can be initially analyzed by conducting descriptive statistics
df_user_id_1.describe()

#### Q11. Check the organization and structure of your notebook. Be sure to include section headings and code comments

#### Q12. Export your df_ords dataframe as “orders_wrangled.csv” in your “Prepared Data” folder.

df_ords_2.to_csv(os.path.join(path, '02 Data','Prepared Data', 'orders_wrangled.csv'), index = False)

#### Q13. Export the df_dep_t_new dataframe as “departments_wrangled.csv” in your “Prepared Data” folder so that you have a “.csv” file of your departments data in the correct format.

df_dep_t_new.to_csv(os.path.join(path, '02 Data','Prepared Data', 'departments_wrangled.csv'), index = False)

# Export to pkl

df_dep_t_new.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'departments_wrangled.pkl'))
