==> Instacart Basket Analysis/02 Scripts/5 Data Consistency Checks.ipynb <==
# Importing libraries

import pandas as pd
import numpy as np
import os

#### Defining path

path = r'C:\Users\nukis\Documents\Instacart Basket Analysis'

### Importing products.csv and orders.csv

df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_wrangled.csv'), index_col= False)

# 2. Data consisteny check steps

# Check for mixed-type data in your df_prods dataframe
for col in df_prods.columns.tolist():
  weird = (df_prods[[col]].applymap(type) != df_prods[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_prods[weird]) > 0:
    print (col)

# Check for missing values
df_prods.isnull().sum()

# View missing values
df_prods_nan = df_prods[df_prods['product_name'].isnull() == True]

df_prods_nan

df_prods_nan['product_name'].dtypes

df_prods_nan.isna().sum()

df_prods.shape

# Copying ONLY non-missing values to new dataframe. In other words, removing missing values.
# Set to False to exclude missing values.
df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]

df_prods_clean.shape

# Finding duplicates
df_dups = df_prods_clean[df_prods_clean.duplicated()]

df_dups

# Addressing duplicates
df_prods_clean_no_dups = df_prods_clean.drop_duplicates()

df_prods_clean_no_dups.shape

# Task 4.5

#### Q2. Run the df.describe() function on your df_prods dataframe. Using your new knowledge about how to interpret the output of this function, share in a markdown cell whether anything about the data looks off or should be investigated further.
Tip: Keep an eye on min and max values!

df_prods.describe()

#### Prices max value is bit weird, as this seems very high, especially knowing that 50th percentile value is 7.1.

#### Q3. Check for mixed-type data in your df_ords dataframe.

# Check for mixed-type data in your df_ords dataframe
for col in df_ords.columns.tolist():
  weird = (df_ords[[col]].applymap(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_ords[weird]) > 0:
    print (col)

#### Q4. If you find mixed-type data, fix it. The column in question should contain observations of a single data type.

#### No mixed-data type is found in the df_ords dataframe.

#### Q5. Run a check for missing values in your df_ords dataframe.
In a markdown cell, report your findings and propose an explanation for any missing values you find.

# Check for missing values
df_ords.isnull().sum()

#### 206.209 missing values from days_since_prior_order.

# View missing values
df_ords_nan = df_ords[df_ords['days_since_prior_order'].isnull() == True]

#### Propose to investigate why those values are missing. My assumption, those values come from first-time customers who don't have previous orders. If that is the case, I propose to fill it with values such as '-1' to flag them and differenciate them with customers who made purchases today (flagged with '0').

#### Q6. Address the missing values using an appropriate method.
In a markdown cell, explain why you used your method of choice.

# Copy the DataFrame
df_ords_clean = df_ords

# Fill the missing values with '-1'.
df_ords_clean['days_since_prior_order'].fillna(-1, inplace = True)

df_ords_clean.head()

df_ords_clean.isnull().sum()

#### As mentioned above, I flagged/filled missing values with '-1'.

#### Q7. Run a check for duplicate values in your df_ords data.
In a markdown cell, report your findings and propose an explanation for any duplicate values you find.

df_ords_clean_dups = df_ords_clean[df_ords_clean.duplicated()]

#### Q8. Address the duplicates using an appropriate method.
In a markdown cell, explain why you used your method of choice.

#### No duplicate is found in the 'df_ords_clean' dataset.

#### Q9. Export your final, cleaned df_prods and df_ords data as “.csv” files in your “Prepared Data” folder and give them appropriate, succinct names.

df_prods_clean_no_dups.to_csv(os.path.join(path, '02 Data','Prepared Data', 'products_cleaned.csv'), index = False)

df_ords_clean.to_csv(os.path.join(path, '02 Data','Prepared Data', 'orders_cleaned.csv'), index = False)
