==> Instacart Basket Analysis/02 Scripts/6 Combining & Exporting Data.ipynb <==
### This script contains the following points:
1. Exporting PKL DataFrame 'orders_products_combined.pkl'
2. Anwers to Task 4.6, No. 3-9

# Import libraries

import pandas as pd
import numpy as np
import os

# Defining path
path = r'C:\Users\nukis\Documents\Instacart Basket Analysis'

#### Q3. In a new notebook, import the orders_products_combined dataframe from the pickle file you just saved.

# Importing 'orders_products_combined.pkl'

df_ords_prods_comb = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_combined.pkl'))

df_ords_prods_comb.head()

#### Q4. Check the shape of the imported dataframe (it should be the same as the one you exported — always check!).

df_ords_prods_comb.shape

#### Yes. It is the same.

#### Q5. Determine a suitable way to combine the 'orders_products_combined' dataframe with your products data set. Make sure you’re using your wrangled, cleaned, and deduped products data set stored in your “Prepared Data” folder from the previous Exercise’s task.

# Importing 'products_cleaned.csv'
df_prods_cleaned = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'products_cleaned.csv'), index_col = False)

df_prods_cleaned.head()

df_prods_cleaned.shape

#### Merging 2 DataFrames: 'orders_products_combined.pkl' and 'products_cleaned.csv'.

# Dropping '_merge' column from 'orders_products_combined.pkl'
df_ords_prods_comb = df_ords_prods_comb.drop(columns = ['_merge'])

df_ords_prods_comb.shape

# Merging
df_ords_prods_merged = df_ords_prods_comb.merge(df_prods_cleaned, on = 'product_id', indicator = True)

#### Q6. Confirm the results of the merge using the merge flag.

df_ords_prods_merged.head()

df_ords_prods_merged['_merge'].value_counts()

df_ords_prods_merged.shape

#### Q7. Export the newly created dataframe as orders_products_merged in a suitable format (taking into consideration the size).

# Import to PKL
df_ords_prods_merged.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_merged.pkl'))
