==> Instacart Basket Analysis/02 Scripts/7 Deriving New Variables.ipynb <==
# This file contains following points:
1. Exercise 4.7
2. Answers Task 4.7

# Import libraries

import pandas as pd
import numpy as np
import os

# Defining path
path = r'C:\Users\nukis\Documents\Instacart Basket Analysis'

# Importing 'orders_products_merged.pkl'

df_orders_products_merged = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_merged.pkl'))

# Dropping '_merge' column 

df_orders_products_merged = df_orders_products_merged.drop(columns = ['_merge'])

# Exercise 4.7

df_orders_products_merged.head()

df = df_orders_products_merged[:1000000]

#### 1. If-Statement using function

# Row is standard argument looking at each row within DataFrame

def price_label(row):

  if row['prices'] <= 5:
    return 'Low-range product'
  elif (row['prices'] > 5) and (row['prices'] <= 15):
    return 'Mid-range product'
  elif row['prices'] > 15:
    return 'High range'
  else: return 'Not enough data'

df['price_range'] = df.apply(price_label, axis=1)

df['price_range'].value_counts(dropna = False)

df['prices'].max()

#### 2. If-statement using loc()

df_orders_products_merged.loc[df_orders_products_merged['prices'] > 15, 'price_range_loc'] = 'High-range product'
df_orders_products_merged.loc[(df_orders_products_merged['prices'] <= 15) & (df_orders_products_merged['prices'] > 5), 'price_range_loc'] = 'Mid-range product'
df_orders_products_merged.loc[df_orders_products_merged['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

df_orders_products_merged['price_range_loc'].value_counts(dropna = False)

#### 3. If-statements using For-Loops

df_orders_products_merged['order_day_of_week'].value_counts(dropna = False)

result = []

for value in df_orders_products_merged["order_day_of_week"]:
  if value == 0:
    result.append("Busiest day")
  elif value == 4:
    result.append("Least busy")
  else:
    result.append("Regularly busy")

df_orders_products_merged['busiest_day'] = result

df_orders_products_merged['busiest_day'].value_counts(dropna = False)

len(result)

df_orders_products_merged.shape

# Task 4.7

#### Q1. If you haven’t done so already, complete the instructions in the Exercise for creating the “price_label” and “busiest_day” columns.
Done

#### Q2. Suppose your clients have changed their minds about the labels you created in your “busiest_day” column. Now, they want “Busiest day” to become “Busiest days” (plural).
This label should correspond with the two busiest days of the week as opposed to the single busiest day. At the same time, they’d also like to know the two slowest days. Create a new column for this using a suitable method.

labeling_counts = df_orders_products_merged['order_day_of_week'].value_counts(dropna = False)

labeling_counts

labeling_counts.index

# As index in this case is day nummber and sorted (ascending), 2 busiest days and 2 slowest days are found.
# I will use it as automated comparison for the days labeling by converting into list.

day_number = list(labeling_counts.index)

day_number

results = []

for value in df_orders_products_merged['order_day_of_week']:
    if value == day_number[0] or value == day_number[1]:
        results.append('Busiest days')
    elif value == day_number[len(day_number) - 1] or value == day_number[len(day_number) - 2]:
        results.append('Slowest days')
    else:
        results.append('Regulary busy days')

# Checking if results length = DataFrame length

len(results)

df_orders_products_merged['order_day_of_week'].shape

# Results length corresponds to the DataFrame. We can just merge it.
# Create a new column in 'df_orders_products_merged' named busiest_days

df_orders_products_merged['busiest_days'] = results

#### Q3. Check the values of this new column for accuracy. Note any observations in markdown format.

df_orders_products_merged['busiest_days'].value_counts(dropna = False)

#### Q4. When too many users make Instacart orders at the same time, the app freezes. The senior technical officer at Instacart wants you to identify the busiest hours of the day. Rather than by hour, they want periods of time labeled “Most orders,” “Average orders,” and “Fewest orders.” Create a new column containing these labels called “busiest_period_of_day.”

# Reminding of how the data looks like

df_orders_products_merged.head(5)

# Converting frequency into dictionary
dict_labeling_counts = df_orders_products_merged['order_hour_of_day'].value_counts(dropna = False).to_dict()

# Checking frequency dictionary
dict_labeling_counts

# Grouping into Most orders, Average orders and Fewest orders, I decided to use percentile as criteria.

list_quantile = list(df_orders_products_merged['order_hour_of_day'].value_counts(dropna = False).quantile([0.75, 0.5, 0.25]))

list_quantile

results_labeling_hour = []

# I used percentiles and nested loop here for grouping automation, 
# instead of hardcoding it using value 'order_hour_of_day' in the If-Statement.

for value in df_orders_products_merged['order_hour_of_day']:
    for key, val in dict_labeling_counts.items():
        if value == key:
            if val < list_quantile[2]:
                results_labeling_hour.append('Fewest orders')
            elif val >= list_quantile[2] and val < list_quantile[1]:
                results_labeling_hour.append('Average orders')
            elif val >= list_quantile[1]:
                results_labeling_hour.append('Most orders')

# Function to check the list before merging it with the DataFrame
def labeling_result (x):
    most_orders = x.count('Most orders')
    average_orders = x.count('Average orders')
    fewest_orders = x.count('Fewest orders')
    
    print(f'Most orders: {most_orders}')
    print(f'Average orders: {average_orders}')
    print(f'Fewest orders: {fewest_orders}')
    print(f'Total: {most_orders + average_orders + fewest_orders}')

# Checking the grouping process through nested loop.
labeling_result(results_labeling_hour)

df_orders_products_merged['order_hour_of_day'].shape

Since both length corresponds to each other, we can integrate it into 'df_orders_products_merged'.

df_orders_products_merged['busiest_period_of_day'] = results_labeling_hour

#### Q5. Print the frequency for this new column.

df_orders_products_merged['busiest_period_of_day'].value_counts(dropna = False)

#### Q6. Ensure your notebook is clean and structured and that your code is well commented.

#### Q7. Export your dataframe as a pickle file (since you added new columns) and store it correctly in your “Prepared Data” folder.

# Export to PKL
df_orders_products_merged.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_merged_update.pkl'))
