==> Instacart Basket Analysis/02 Scripts/9 Intro to Data Visualization with Python Part1A.ipynb <==
## This file contains anwers Task 4.9 Part 1

# Import libraries

import pandas as pd
import numpy as np
import os
import gc

# Defining path
path = r'C:\Users\nukis\Documents\Instacart Basket Analysis'

# Importing 'orders_products_merged.pkl'

df_cust = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'customers.csv'))

#### Q4. Wrangle the data so that it follows consistent logic.

df_cust.head(10)

df_cust.tail(10)

df_cust.shape

# Renaming columns
df_cust.rename(columns= {'First Name' : 'first_name', 'Surnam' : 'surname','Gender' : 'gender','STATE' : 'state',
                          'Age' : 'age', 'n_dependants': 'dependants_number', 'fam_status' : 'marital_status'}, inplace = True)

df_cust.head(5)

No columns being dropped, as they all relevant for the anaysis.

#### Q5. Complete the fundamental data quality and consistency checks.

# Check for missing values

df_cust.isnull().sum()

There are 11,259 missing values found in the 'first_name' column.

# View missing values

df_cust[df_cust['first_name'].isnull() == True]

1. As the missing values are only 'first_name', it does not really affect our analysis. It should be considered that firstly, person name should remain confidential/privacy (unless we need it for loyalty program) and secondly, we still have acess to these user through their 'user_id' and 'surname'.
2. So, I came up with solution to fill these missing values to flag this as follows:
- If the gender is female, it will be filled with Ms.
- If the gender is male, it will be filled with Mr.

# Copy the DataFrame

df_cust_clean = df_cust

# Only as note. This command resulted in warning. Thus, it was not used.
# The next command run perfectly, no warning.

df_cust_clean['first_name'] = np.where((df_cust_clean['first_name'].isnull()) & (df_cust_clean['gender'] == 'Female'), 'Ms', df_cust_clean['first_name'])

df_cust_clean.loc[(df_cust_clean['first_name'].isnull()) & (df_cust_clean['gender'] == 'Female'), 'first_name'] = 'Ms'
df_cust_clean.loc[(df_cust_clean['first_name'].isnull()) & (df_cust_clean['gender'] == 'Male'), 'first_name'] = 'Mr'

# Check again if those above command worked.

df_cust.isnull().sum()

Ms = len(df_cust_clean[df_cust_clean['first_name'] == 'Ms'].value_counts(dropna = False))
Mr = len(df_cust_clean[df_cust_clean['first_name'] == 'Mr'].value_counts(dropna = False))
print(f'Ms as first_name: {Ms}')
print(f'Mr as first_name: {Mr}')
print(f'Total: {Ms + Mr}')

# Check for mixed-type data in 'df_cust_clean'.

for col in df_cust_clean.columns.tolist():
  weird = (df_cust_clean[[col]].applymap(type) != df_cust_clean[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_cust_clean[weird]) > 0:
    print (col)

df_cust_clean.dtypes

No mixed-data type is found in the 'df_cust_clean' dataframe.

# Check for duplicates

df_cust_clean[df_cust_clean.duplicated()]

No duplicates are found in the 'df_cust_clean' dataframe.

# Export the clean customer DataFrame to avoid Jupyter crash.

df_cust_clean.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'customer_cleaned.pkl'))

gc.collect()

#### Q6. Combine your customer data with the rest of your prepared Instacart data.

As my Jupyter kept crashing, I divided this notebook into 3 parts. This notebook will be continued to Part1B and Part1C.
