==> Instacart Basket Analysis/02 Scripts/9 Intro to Data Visualization with Python Part1B.ipynb <==
## This file contains anwers Task 4.9 Part 1

# Import libraries

import pandas as pd
import numpy as np
import os
import gc

# Defining path
path = r'C:\Users\nukis\Documents\Instacart Basket Analysis'

# Importing 'orders_products_merged_update2.pkl'

df_ords_prods_cleaned = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_merged_update2.pkl'))

#### Datatype Conversions

df_ords_prods_cleaned.info(memory_usage='deep')

# Converting object data type to category for all columns below except 'product_name' as it has too many unique values.

df_ords_prods_cleaned_obj = df_ords_prods_cleaned.select_dtypes(include=['object']).copy()
df_ords_prods_cleaned_obj.describe()

# Change in datatype by downcasting loop

for column in df_ords_prods_cleaned:
    if df_ords_prods_cleaned[column].dtype == 'float64':
        df_ords_prods_cleaned[column] = pd.to_numeric(df_ords_prods_cleaned[column], downcast = 'float')
    elif df_ords_prods_cleaned[column].dtype == 'int64':
        df_ords_prods_cleaned[column] = pd.to_numeric(df_ords_prods_cleaned[column], downcast = 'integer')
    else:
        if df_ords_prods_cleaned[column].dtype == 'object' and column != 'product_name':
            df_ords_prods_cleaned[column] = df_ords_prods_cleaned[column].astype('category') 

df_ords_prods_cleaned.info(memory_usage='deep')

By using above steps, memory usage is reduced from 21 GB to 4.2 GB to prevent my Jupyter from crashing all the time.

# Export the resized version for emergency situation in case it crashes again.

df_ords_prods_cleaned.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_merged_update2_resized.pkl'))

gc.collect()

As my Jupyter kept crashing, I restarted kernel and cleared the output. This notebook will be continued to Part1C.
