==> Instacart Basket Analysis/02 Scripts/9 Intro to Data Visualization with Python Part1C.ipynb <==
## This file contains anwers Task 4.9 Part 1

# Import libraries

import pandas as pd
import numpy as np
import os
import gc

# Defining path
path = r'C:\Users\nukis\Documents\Instacart Basket Analysis'

# Importing resized version of 'orders_products_merged_update2.pkl'

df_ords_prods_cleaned = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_merged_update2_resized.pkl'))

# Import again the cleaned version of 'df_cust_clean'

df_cust_cleaned = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'customer_cleaned.pkl'))

df_cust_cleaned.info(memory_usage='deep')

# Looking at the size of this DataFrame, I think, reducing 'object' datatype is not really that relevant.

df_cust_cleaned_obj = df_cust_cleaned.select_dtypes(include=['object']).copy()
df_cust_cleaned_obj.describe()

# Change in datatype by downcasting loop

for column in df_cust_cleaned:
    if df_cust_cleaned[column].dtype == 'int64':
        df_cust_cleaned[column] = pd.to_numeric(df_cust_cleaned[column], downcast = 'integer')

df_cust_cleaned.info(memory_usage='deep')

df_cust_cleaned.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'customer_cleaned_resized.pkl'))

# Merging resized 'df_ords_prods_clean' and 'df_cust_clean'. The data type 'user_id' of both DataFrame is already the same.

df_ords_prods_cust_cleaned = df_ords_prods_cleaned.merge(df_cust_cleaned, on = 'user_id', indicator = True)

# Export the merged 'df_ords_prods_cleaned' and 'df_cust_cleaned'

df_ords_prods_cust_cleaned.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'ords_prods_cust_cleaned.pkl'))

df_ords_prods_cust_cleaned.head(10)

df_ords_prods_cust_cleaned['_merge'].value_counts()

df_ords_prods_cust_cleaned.shape

df_ords_prods_cust_cleaned.dtypes

df_ords_prods_cust_cleaned.info(memory_usage='deep')

gc.collect()
